+++ src/transpiler.ts
@@ -9,7 +9,7 @@ import {
   ExpressionSplitter,
   ExternalInputChecker,
   ExternImporter,
-  FreeFunctions,
+  FreeLibraryCallInliner,
   IdentifierMangler,
   IfFunctionaliser,
   ImplicitConversionToExplicit,
@@ -81,7 +81,7 @@ function applyPasses(ast: AST, options: TranspilationOptions): AST {
     ['Ib', IntBoundCalculator],
     ['Ch', ConstantHandler],
     ['M', IdentifierMangler],
-    ['Ff', FreeFunctions],
+    ['Fi', FreeLibraryCallInliner],
     ['Rl', ReferencedLibraries],
     ['Ii', InheritanceInliner],
     ['Mh', ModifierHandler],
